common: move single-use function MoveNewFilesTo to pkg/endpoint
This function is only used in pkg/endpoint, so move it there and
unexport it.
